Added dashboard path and base landing page
