+++ src/routes/+page.server.ts
@@ -1,21 +1,6 @@
-import { redirect, type Actions, fail } from '@sveltejs/kit';
-import { auth } from '$lib/server';
+import { redirect } from '@sveltejs/kit';
 
 export const load = async ({ locals }) => {
 	const { user } = await locals.auth.validateUser();
-	if (!user) throw redirect(302, '/login');
-
-	return {
-		user
-	};
-};
-
-export const actions: Actions = {
-	// signout
-	default: async ({ locals }) => {
-		const session = await locals.auth.validate();
-		if (!session) return fail(401);
-		await auth.invalidateSession(session.sessionId);
-		locals.auth.setSession(null);
-	}
+	if (user) throw redirect(302, '/dashboard');
 };
