+++ src/routes/+page.svelte
@@ -1,22 +1,7 @@
 <script lang="ts">
-	import { enhance } from '$app/forms';
-	import { t, locale, locales, translate } from '$locales';
-
-	export let data;
+	import { t } from '$locales';
 </script>
 
-<p>{$t('herotitle')}</p>
-<pre class="code">
-{JSON.stringify(data.user, null, 2)}
-</pre>
-<a href="/live">{$t('getstarted')}</a>
-
-<select bind:value={$locale}>
-	{#each locales as l}
-		<option value={l}>{translate(l, 'locale')}</option>
-	{/each}
-</select>
-
-<form use:enhance method="post">
-	<input type="submit" class="button" value={$t('signout')} />
-</form>
+<h1>{$t('herotitle')}</h1>
+<a href="/login">{$t('login')}</a>
+<a href="/signup">{$t('signup')}</a>
